update test filters and add test for op registration check
